+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased yet
+## 1.4.11
 
 ADDED:
 - get-payload: Add support for reading files from Multipart Form Requests.